logout router added

+++ src/components/CustomerDashboard.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useNavigate } from "react-router-dom";
 import '../App.css';
+import {jwtDecode} from 'jwt-decode';
 const CustomerDashboard = () => {
   const[token,setToken] = useState('')
 
@@ -8,30 +9,50 @@ const CustomerDashboard = () => {
 
 
 
-      const handleLogout = ()=>{
-      if( localStorage.getItem('authToken') !== null) {
-        navigation('/')
-
-           localStorage.clear()
+  const handleLogout = async () => {
+    if (localStorage.getItem('authToken') !== null) {
+      const response = await fetch('https://brand-wick.onrender.com/user/logout', {
+        method: 'POST',
+        headers: {
+          'Authorization': `Bearer ${localStorage.getItem('authToken')}`, 
+          'Content-Type': 'application/json',
+        },
+      });
+  
+      if (response.ok) {
+        localStorage.removeItem('authToken');
+        localStorage.clear();
+        console.log('Successfully logged out',token)
+        navigation('/');
+      } else {
+        console.error('Logout failed');
       }
-      
     }
+  };
+  
 
-    useEffect(()=>{
-      
-        const check = () => {
-          let homeCheck = window.location.href.split("/")
-        if ( localStorage.getItem('authToken') === null && homeCheck[3] === "dashboard"){
-          navigation('/')
-      }
-      else if(localStorage.getItem("authToken")) {
-        console.log()
-         setToken(localStorage.getItem("authToken"))
-      }
-          }
+  useEffect(() => {
+    const checkToken = () => {
+      let homeCheck = window.location.href.split("/");
+      if (localStorage.getItem('authToken') === null && homeCheck[3] === "dashboard") {
+        navigation('/');
+      } else if (localStorage.getItem("authToken")) {
+        const decodedToken = jwtDecode(localStorage.getItem("authToken"));
+        const expirationTime = decodedToken.exp * 1000; 
+        const currentTime = Date.now();
 
-    check()
-    },[]) 
+        if (currentTime >= expirationTime) {
+          console.log('Token has expired');
+          localStorage.removeItem('authToken');
+          navigation('/');
+        } else {
+          setToken(localStorage.getItem("authToken"));
+        }
+      }
+    };
+
+    checkToken();
+  }, [navigation]);
   
 
 
@@ -42,7 +63,8 @@ const CustomerDashboard = () => {
      
       <div className='grid'>
       
-       <h3>Token:</h3><span>{token}</span>
+       <h3>Token:</h3>
+       <p>{token}</p>
       </div>
       
     </div>
